Refactor adminUsers hooks
